refactor: 유틸리티 response 함수를 response 클래스 멤버함수로 대체
