+++ default.conf
@@ -11,6 +11,7 @@ http {
         client_max_body_size 20m;
 
         location / {
+			index index.html;
             root ./var/www/main;
             allow_methods GET POST;
         }
