+++ src/handler/cgi/Executor.hpp
@@ -13,7 +13,6 @@
 #include "../../router/model/RouteDecision.hpp"
 #include "../../server/epoll/manager/EpollManager.hpp"
 #include "../exception/Exception.hpp"
-#include "../utils/response.hpp"
 #include "ProcessManager.hpp"
 
 namespace handler {
